feat: split page html using components
