+++ packages/core/src/components/account/components/MyAccountStatusBadge/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MyAccountStatusBadge'
