+++ packages/core/src/components/account/components/MyAccountStatusBadge/styles.module.scss
@@ -1,0 +1,27 @@
+.status {
+  display: inline-block;
+  width: fit-content;
+  height: fit-content;
+  padding: var(--fs-spacing-0) var(--fs-spacing-2);
+  overflow: hidden;
+  line-height: 1;
+  color: var(--fs-color-text);
+  white-space: nowrap;
+  border: 0 none transparent;
+  border-radius: var(--fs-border-radius-pill);
+}
+
+.statusSuccess {
+  background-color: #e9fce3;
+  border-color: #e9fce3;
+}
+
+.statusWarning {
+  background-color: #ffedcd;
+  border-color: #ffedcd;
+}
+
+.statusNeutral {
+  background-color: #f5f5f5;
+  border-color: #f5f5f5;
+}
