+++ packages/core/src/components/account/orders/MyAccountListOrders/MyAccountListOrdersTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MyAccountListOrdersTable'
