+++ packages/core/src/components/account/orders/MyAccountListOrders/index.ts
@@ -1,0 +1,1 @@
+export { default as MyAccountListOrders } from './MyAccountListOrders'
